+++ src/main/groovy/com/rapidminer/gradle/RapidMinerExtensionPlugin.groovy
@@ -15,6 +15,7 @@
  */
 package com.rapidminer.gradle
 
+import org.gradle.api.GradleException
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.file.FileTree
@@ -66,7 +67,7 @@ class RapidMinerExtensionPlugin implements Plugin<Project> {
 			apply plugin: 'maven-publish'
 			apply plugin: 'java'
 			apply plugin: 'com.github.johnrengelman.shadow'
-			
+
 			ext {
 				// Publishing variables
 				SNAPSHOT = '-SNAPSHOT'
@@ -75,7 +76,7 @@ class RapidMinerExtensionPlugin implements Plugin<Project> {
 				RELEASE_REPO = 'libs-release-local'
 				REPO = version.endsWith(SNAPSHOT) ?  SNAPSHOT_REPO : RELEASE_REPO
 			}
-			
+
 			// define Maven publications
 			publishing {
 				publications {
@@ -99,7 +100,12 @@ class RapidMinerExtensionPlugin implements Plugin<Project> {
 			// shadowJar is being used to create a shaded extension jar
 			apply plugin: 'com.rapidminer.gradle.java-basics'
 			apply plugin: 'com.rapidminer.gradle.code-quality'
-			apply plugin: 'com.rapidminer.gradle.release'
+
+			try {
+				apply plugin: 'com.rapidminer.gradle.release'
+			} catch(e) {
+				project.logger.error "Could not apply release plugin", e
+			}
 
 			defaultTasks 'installExtension'
 
@@ -117,11 +123,6 @@ class RapidMinerExtensionPlugin implements Plugin<Project> {
 				from shadowJar
 			}
 
-			// create publish all task
-			def publishLibAndRelease = tasks.create(name: 'publishExtension', dependsOn: 'publishExtensionJarPublicationToMavenRepository')
-			publishLibAndRelease.group = EXTENSION_GROUP
-			publishLibAndRelease.description = "Publishes extension .jar and shaded extension .jar to configured Maven repository."
-
 			// add and configure Gradle wrapper task
 			tasks.create(name: 'wrapper', type: org.gradle.api.tasks.wrapper.Wrapper)
 			wrapper { gradleVersion = "${->extensionConfig.wrapperVersion}" }
@@ -130,36 +131,41 @@ class RapidMinerExtensionPlugin implements Plugin<Project> {
 			// see http://forums.gradle.org/gradle/topics/how_do_you_delay_configuration_of_a_task_by_a_custom_plugin_using_the_extension_method
 			group = "${->project.extensionConfig.groupId}"
 
+			def checkManifestTask = tasks.create(name: 'checkManifestEntries')
+			checkManifestTask.group = EXTENSION_GROUP
+			checkManifestTask.description = "Checks whether the extension manifest entries are correct."
+
+			checkManifestEntries {
+				doLast {
+					if(!project.extensionConfig.name){
+						throw new GradleException('No RapidMiner Extension name defined. Define via \'extensionConfig { name $NAME }\'.')
+					}
+
+					if(!project.extensionConfig.groupId) {
+						throw new GradleException('No groupdId defined! Define via \'extensionConfig { groupdId $GROUPID }\'. (default: \'com.rapidminer.extension\')')
+					}
+					checkReleaseManifestEntries(project)
+				}
+			}
+			jar.dependsOn checkManifestEntries
+			shadowJar.dependsOn checkManifestEntries
+
 			// Configuring the properties below can only be accomplished after
 			// the project extension 'extension' has been configured
 			afterEvaluate {
 
-				//TODO do these sanity checks only when jar is created (as import of project without an extension name does not work)
-				// first sanity checks
-				if(!project.extensionConfig.name){
-					throw new RuntimeException("No RapidMiner Extension name defined. Define via 'extensionConfig { name $NAME }'.")
-				}
-
-				if(!project.extensionConfig.groupId) {
-					throw new RuntimeException("No groupdId defined! Define via 'extensionConfig { groupdId $GROUPID }'. (default: 'com.rapidminer.extension')")
-				}
-
 				// add RapidMiner and configured extensions as dependency to all projects
 				allprojects { Project p ->
-					if(p.plugins.hasPlugin(JavaPlugin)) {
-						dependencies {
-							provided getRapidMinerDependency(project)
-							extensionConfig.dependencies.extensions.each{  e ->
-								provided group: e.group, name: e.namespace, version: e.version
-							}
+					dependencies {
+						provided getRapidMinerDependency(project)
+						extensionConfig.dependencies.extensions.each{  e ->
+							provided group: e.group, name: e.namespace, version: e.version
 						}
 					}
 				}
 
-				// TODO create tasks that checks entries instead of doing it in the evaluation phase
-				// TODO jar should depend on this new task
 				// add check for manifest entries to avoid generic 'null' error
-				checkReleaseManifestEntries(project)
+				assignReleaseManifestEntries(project)
 
 				// configure create extension release task
 				jar {
@@ -167,29 +173,29 @@ class RapidMinerExtensionPlugin implements Plugin<Project> {
 					manifest {
 						attributes(
 								"Manifest-Version": 		"1.0",
-								"Implementation-Vendor": 	project.extensionConfig.vendor,
-								"Implementation-Title":		project.extensionConfig.name,
-								"Implementation-URL":		project.extensionConfig.homepage,
-								"Implementation-Version": 	project.version,
-								"Specification-Title": 		project.extensionConfig.name,
-								"Specification-Version":	project.version,
-								"RapidMiner-Version":		getRapidMinerVersion(project),
+								"Implementation-Vendor": 	project.extensionConfig.vendor ?: '',
+								"Implementation-Title":		project.extensionConfig.name ?: '',
+								"Implementation-URL":		project.extensionConfig.homepage ?: '',
+								"Implementation-Version": 	project.version ?: '',
+								"Specification-Title": 		project.extensionConfig.name ?: '',
+								"Specification-Version":	project.version ?: '',
+								"RapidMiner-Version":		getRapidMinerVersion(project) ?: '',
 								"RapidMiner-Type":			"RapidMiner_Extension",
-								"Plugin-Dependencies":		getExtensionDependencies(project),
+								"Plugin-Dependencies":		getExtensionDependencies(project) ?: '',
 
 								// Definition of important files
-								"Extension-ID":					RMX + project.extensionConfig.namespace,
-								"Namespace":					project.extensionConfig.namespace,
-								"Initialization-Class":			project.extensionConfig.resources.initClass,
-								"IOObject-Descriptor":			project.extensionConfig.resources.objectDefinition,
-								"Operator-Descriptor":			project.extensionConfig.resources.operatorDefinition,
-								"ParseRule-Descriptor":			project.extensionConfig.resources.parseRuleDefinition,
-								"Group-Descriptor":				project.extensionConfig.resources.groupProperties,
-								"Error-Descriptor":				project.extensionConfig.resources.errorDescription,
-								"UserError-Descriptor":			project.extensionConfig.resources.userErrors,
-								"GUI-Descriptor":				project.extensionConfig.resources.guiDescription,
-								"Settings-Descriptor":			project.extensionConfig.resources.settingsDescriptor,
-								"SettingsStructure-Descriptor":	project.extensionConfig.resources.settingsStructureDescriptor
+								"Extension-ID":					RMX + (project.extensionConfig.namespace ?: ''),
+								"Namespace":					project.extensionConfig.namespace ?: '',
+								"Initialization-Class":			project.extensionConfig.resources.initClass ?: '',
+								"IOObject-Descriptor":			project.extensionConfig.resources.objectDefinition ?: '',
+								"Operator-Descriptor":			project.extensionConfig.resources.operatorDefinition ?: '',
+								"ParseRule-Descriptor":			project.extensionConfig.resources.parseRuleDefinition ?: '',
+								"Group-Descriptor":				project.extensionConfig.resources.groupProperties ?: '',
+								"Error-Descriptor":				project.extensionConfig.resources.errorDescription ?: '',
+								"UserError-Descriptor":			project.extensionConfig.resources.userErrors ?: '',
+								"GUI-Descriptor":				project.extensionConfig.resources.guiDescription ?: '',
+								"Settings-Descriptor":			project.extensionConfig.resources.settingsDescriptor ?: '',
+								"SettingsStructure-Descriptor":	project.extensionConfig.resources.settingsStructureDescriptor ?: ''
 								)
 					}
 				}
@@ -208,40 +214,66 @@ class RapidMinerExtensionPlugin implements Plugin<Project> {
 		return project.extensionConfig.namespace
 	}
 
-	/**
-	 * Checks for existence of user defined resource objects. If no objects are defined by the user
-	 *
-	 */
 	def checkReleaseManifestEntries(Project project) {
 		assert project.version
 		assert project.extensionConfig.vendor
 
 		def res = project.extensionConfig.resources
-		def name = project.extensionConfig.name.replace(" ", "")
+		def name = project.extensionConfig.name?.replace(" ", "") ?: ''
 		def logger = project.logger
 
-		res.initClass = checkInitClass(project, res, name, logger)
-		res.operatorDefinition = checkResourceFile("Operators", XML_EXTENSION, res.operatorDefinition, project, res, name, logger, true)
-		res.objectDefinition = checkResourceFile("ioobjects", XML_EXTENSION, res.objectDefinition, project, res, name, logger)
-		res.parseRuleDefinition = checkResourceFile("parserules", XML_EXTENSION, res.parseRuleDefinition, project, res, name, logger)
-		res.groupProperties = checkResourceFile("groups", PROPERTIES_EXTENSION, res.groupProperties, project, res, name, logger)
-		res.errorDescription = checkResourceFile("Errors", PROPERTIES_EXTENSION, res.errorDescription, project, res, name, logger, false, I18N_PATH)
-		res.userErrors = checkResourceFile("UserErrorMessage", PROPERTIES_EXTENSION, res.userErrors, project, res, name, logger, false,  I18N_PATH)
-		res.guiDescription = checkResourceFile("GUI", PROPERTIES_EXTENSION, res.guiDescription, project, res, name, logger,  false, I18N_PATH)
-		res.settingsStructureDescriptor = checkResourceFile("settings", XML_EXTENSION, res.settingsStructureDescriptor, project, res, name, logger,  false)
-		res.settingsDescriptor = checkResourceFile("Settings", PROPERTIES_EXTENSION, res.settingsDescriptor, project, res, name, logger,  false, I18N_PATH)
+		// Check for Initialization class
+		checkInitClass(project, res, name, logger)
+
+		// Check for Operator definitions
+		def operatorsResource = checkResourceFile("Operators", XML_EXTENSION, res.operatorDefinition, project, res, name, logger, true)
+		def resourceFile = project.file(DEFAULT_RESOURCE_PATH + operatorsResource)
+		def docBundle = new XmlSlurper().parse(resourceFile)?.attributes()?.get('docbundle')
+		if(!docBundle){
+			throw new GradleException("No docBundle in operators definitions '${operatorsResource}' defined. \nWrong resource file selected? Please define path to Operator definitions manually.\nDefault path to operator definitions: ${RESOURCE_PACKAGE}Operators${name}.xml")
+		}
+
+		def docBundleFile = project.file(DEFAULT_RESOURCE_PATH + docBundle + ".xml")
+		if(!docBundleFile.exists()){
+			throw new GradleException("DocBundle file defined in operators definitions ('${docBundle}') does not exists.")
+		}
+	}
+
+
+	def assignReleaseManifestEntries(Project project) {
+		def res = project.extensionConfig.resources
+		def name = project.extensionConfig.name?.replace(" ", "") ?: ''
+		def logger = project.logger
+
+		res.initClass = getInitClass(project, res, name, logger)
+		res.operatorDefinition = getResourceFile("Operators", XML_EXTENSION, res.operatorDefinition, project, res, name, logger, true)
+		res.objectDefinition = getResourceFile("ioobjects", XML_EXTENSION, res.objectDefinition, project, res, name, logger)
+		res.parseRuleDefinition = getResourceFile("parserules", XML_EXTENSION, res.parseRuleDefinition, project, res, name, logger)
+		res.groupProperties = getResourceFile("groups", PROPERTIES_EXTENSION, res.groupProperties, project, res, name, logger)
+		res.errorDescription = getResourceFile("Errors", PROPERTIES_EXTENSION, res.errorDescription, project, res, name, logger, false, I18N_PATH)
+		res.userErrors = getResourceFile("UserErrorMessage", PROPERTIES_EXTENSION, res.userErrors, project, res, name, logger, false,  I18N_PATH)
+		res.guiDescription = getResourceFile("GUI", PROPERTIES_EXTENSION, res.guiDescription, project, res, name, logger,  false, I18N_PATH)
+		res.settingsStructureDescriptor = getResourceFile("settings", XML_EXTENSION, res.settingsStructureDescriptor, project, res, name, logger,  false)
+		res.settingsDescriptor = getResourceFile("Settings", PROPERTIES_EXTENSION, res.settingsDescriptor, project, res, name, logger,  false, I18N_PATH)
+	}
+
+	def getInitClass(Project project, res, name, logger){
+		try {
+			return checkInitClass(project, res, name, logger)
+		} catch(e) {
+			return "\$NOT_FOUND\$"
+		}
 	}
 
 	def checkInitClass(Project project, res, name, logger){
 		// Check if init class is user defined
+		def defaultFileName = DEFAULT_JAVA_PATH + RAPIDMINER_PACKAGE + INIT_CLASS_PREFIX + name + JAVA_EXTENSION
 		if(!res.initClass){
-			def defaultFileName = DEFAULT_JAVA_PATH + RAPIDMINER_PACKAGE + INIT_CLASS_PREFIX + name + JAVA_EXTENSION
 			if(project.file(defaultFileName).exists()){
-				logger.info("Found default init class: '" + defaultFileName + "'")
+				logger.info "Found default init class: '${defaultFileName}'"
 				return RAPIDMINER_PACKAGE.replace("/", ".") + INIT_CLASS_PREFIX + name
 			} else {
-				logger.info("Default init class  '" + defaultFileName + "' not found. "
-						+"Searching for alternatives in src/main/java...")
+				logger.info("Default init class  '${defaultFileName}' not found. Searching for alternatives in 'src/main/java' ...")
 
 				// Create a file tree with a base directory
 				FileTree tree = project.fileTree(dir: DEFAULT_JAVA_PATH, include: '**/*' + JAVA_EXTENSION)
@@ -250,7 +282,7 @@ class RapidMinerExtensionPlugin implements Plugin<Project> {
 				def initCandidate = null
 				tree.find { File file ->
 					if(file.getName().contains(INIT_CLASS_PREFIX)){
-						logger.info("Found potential init class: " + file.getPath())
+						logger.info("Found potential init class: ${file.getPath()}")
 						def idx = file.getPath().indexOf(DEFAULT_JAVA_PATH.replace("/", File.separator))
 						initCandidate = file.getPath()
 								.substring(idx + DEFAULT_JAVA_PATH.length())
@@ -263,26 +295,35 @@ class RapidMinerExtensionPlugin implements Plugin<Project> {
 
 				// Still not found?
 				if(initCandidate == null){
-					throw new RuntimeException("No init class candidate found!")
+					throw new GradleException("No init class candidate found!")
 				}
-				logger.info("Selected init class: '"+ initCandidate+"'")
+				logger.info "Selected init class: '${initCandidate}'"
 				return initCandidate
 			}
 		} else {
 			// check if user defined init class exists
 			def initClassFile = project.file(DEFAULT_JAVA_PATH + res.initClass.replace(".", File.separator) + ".java")
 			if(!initClassFile.exists()){
-				throw new RuntimeException("Defined extension init class '"+ initClassFile +"' does not exist!")
+				throw new GradleException("Extension init class '${initClassFile}' does not exist! \nDefault path: ${defaultFileName}")
 			}
 			return res.initClass // use the user-defined one
 		}
 	}
 
+	def getResourceFile(String resourceName, String suffix, userDefinedResource, Project project, res, name, logger, boolean mandatory = false, String subdirectory = ""){
+		try {
+			return checkResourceFile(resourceName, suffix, userDefinedResource, project, res, name, logger, mandatory, subdirectory)
+		} catch(e){
+			e.printStackTrace()
+			logger.info "Could not find resource file for resource '${resourceName}'"
+			return "\$NOT_FOUND\$"
+		}
+	}
 
 	def checkResourceFile(String resourceName, String suffix, userDefinedResource, Project project, res, name, logger, boolean mandatory = false, String subdirectory = ""){
 		// Check if resource file is user defined
+		def defaultResourceFile = RESOURCE_PACKAGE + subdirectory + resourceName + name + suffix
 		if(!userDefinedResource){
-			def defaultResourceFile = RESOURCE_PACKAGE + subdirectory + resourceName + name + suffix
 			if(project.file(DEFAULT_RESOURCE_PATH + defaultResourceFile).exists()){
 				logger.info("Found default " + resourceName + " resource file: '" + defaultResourceFile + "'")
 				return defaultResourceFile
@@ -291,7 +332,7 @@ class RapidMinerExtensionPlugin implements Plugin<Project> {
 						+" Searching for alternatives in "+ DEFAULT_RESOURCE_PATH + "...")
 
 				// Create a file tree with a base directry
-				FileTree tree = project.fileTree(dir: DEFAULT_RESOURCE_PATH, include: '**/*' + suffix)
+				FileTree tree = project.fileTree(dir: DEFAULT_RESOURCE_PATH, include: "**/*${suffix}")
 
 				// Iterate over the contents of a tree
 				def resourceCandidate = null
@@ -310,13 +351,13 @@ class RapidMinerExtensionPlugin implements Plugin<Project> {
 				// Still not found?
 				if(resourceCandidate == null){
 					if(mandatory){
-						throw new RuntimeException("Mandatory " + resourceName + " resource file is missing. No candidate found!")
+						throw new GradleException("Mandatory '${resourceName}' resource file is missing. No candidate found!")
 					} else {
 						resourceCandidate = ""
 					}
-					logger.info("No optional " + resourceName + " resource file found. Skipping...")
+					logger.info("No optional resource file for '${resourceName}' found. Skipping...")
 				} else {
-					logger.info("Selected " + resourceName + " resource file: '"+ resourceCandidate+"'")
+					logger.info("Selected resource file for '${resourceName}': ${resourceCandidate}")
 				}
 				return resourceCandidate
 			}
@@ -324,8 +365,12 @@ class RapidMinerExtensionPlugin implements Plugin<Project> {
 			def fileName = DEFAULT_RESOURCE_PATH + userDefinedResource
 			def resourceFile = project.file(fileName)
 			if(!resourceFile.exists()){
-				throw new RuntimeException("Selected " + resourceName + " resource file '"+ resourceFile +"' does not exist!")
+				throw new GradleException("Resource file for resource '${resourceName}' does not exist! \nDefault path: ${defaultResourceFile}")
 			}
+			def idx = resourceFile.getPath().indexOf(DEFAULT_RESOURCE_PATH.replace("/", File.separator))
+			return resourceFile.getPath()
+					.substring(idx + DEFAULT_RESOURCE_PATH.length())
+					.replace(File.separator, "/")
 		}
 	}